working on C++ file arrangements
